fixed more tests

+++ src/Forms/src/Xamarin.Forms.Core/Shapes/EllipseGeometry.cs
@@ -20,10 +20,10 @@ namespace Xamarin.Forms.Shapes
 			BindableProperty.Create(nameof(Center), typeof(PointF), typeof(EllipseGeometry), new PointF());
 
 		public static readonly BindableProperty RadiusXProperty =
-			BindableProperty.Create(nameof(RadiusX), typeof(double), typeof(EllipseGeometry), 0.0);
+			BindableProperty.Create(nameof(RadiusX), typeof(float), typeof(EllipseGeometry), 0f);
 
 		public static readonly BindableProperty RadiusYProperty =
-			BindableProperty.Create(nameof(RadiusY), typeof(double), typeof(EllipseGeometry), 0.0);
+			BindableProperty.Create(nameof(RadiusY), typeof(float), typeof(EllipseGeometry), 0f);
 
 		public PointF Center
 		{
